permisos para usar recursos

+++ app/Http/Controllers/TareaController.php
@@ -3,16 +3,19 @@
 namespace App\Http\Controllers;
 
 use App\Models\Tarea;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class TareaController extends Controller{
 
     public function index(){
-        return Tarea::all();
+        //return Tarea::all();
+       $tareas = Tarea::where('user_id', auth()->user()->id)->get();
+        return $tareas;
     }
 
     public function show(Tarea $tarea){
-        return $tarea;
+        return $tareas;
     }
 
     public function store(Request $request){
@@ -24,14 +27,14 @@ class TareaController extends Controller{
     }
 
     public function update(Request $request, Tarea $tarea){
+        $this->authorize('update',$tarea);
         $tarea->update($request->all());
-        return response()->json($tarea,200);
+        return response()->json($tarea, 200);;
     }
 
     public function delete(Request $request, Tarea $tarea){
+        $this->authorize('delete',$tarea);
         $tarea->delete();
         return response()->json($tarea,204);
     }
-
-
 }
